Deleted hipchat_notifier.py, Add slack_notifier.py and Integrated slack into shippable.yml
