+++ shippable.yml
@@ -7,9 +7,9 @@ node_js:
 build_image: shippableimages/ubuntu1404_nodejs
 env: 
   global:
-    - PROJECT=Dashboard HIPCHAT_ROOM=announce
+    - SLACK_ORG=Ottemo Project=Dashboard
     - REMOTE_HOST=admin.blitz.ottemo.io FOUNDATION_URI=https://api.blitz.ottemo.io
-    - secure: JePvX8e49vPPWRLagd9lvuwV8wRBUD1EDkHnANJ8bTzfnDiAPBgqrgocliz7KR6F5ZPHBAvRkrIa6ZP1PVLdC/LeOzoojFyPzU7iiKxkatT2+Ry+QKnH603UGJzIfui1bLQ/eZT2D8wOjMggs4z8DXH62HCmzl+ovhboU1XBYRBuYVMZBKKN/2kcz8LVTZ0BP9jLLMJ5xHb9PDOX7fwBREnjn0dN2yy6uNpz9t7rKJCxNf4uk7+M6AHrryU4mEHpEfu9wbUhMGve5Del78YcZfc5PMDtmKqvamUDICOMCzHRFdl3kpkNsr72ynoZ3s0Vus/3cH9hShCfHNacVk4xQA==
+    - secure: E5KMgEfDJ7QcMF4vEG3eiayrs+BUI/EQvrewOW66D3ZIy0cB5xZYrsSsft4TVxr3os+1geUmLMjL+g/O8Sg89o+yG5qlbdodOLBT9qRXjlGEAVh3dlJuTSOFV0t4hNIP8vBtdgitLKQwzg3QBmJRjjEh734vjjT7SnS6jXiCHb2s81H+QD0dZTnIgg3b3m2nbKpm5TfgCjI0t4810+zPU/L3uV7mWbX6s77iWlPafIlYgXZxY7hMViSavlbbLwTAr3T08iu7mB/3Ti2CxxI6MQB9jxFiFP9yvKKVmF9plmpjqzumZ/Pd3q6wxBdk6gkLmGWyDWM/mtpanp9zTW2DNQ==
 
 before_install:
     # install NodeJS v0.12
@@ -28,7 +28,7 @@ script:
     - bash deploy.sh
 
 after_failure:
-    - python hipchat_notifier.py --project $PROJECT --room $HIPCHAT_ROOM --token $HIPCHAT_TOKEN
+    - python slack_notifier.py --org $SLACK_ORG --project $PROJECT --token $SLACK_TOKEN
 
 after_success:
-    - python hipchat_notifier.py --project $PROJECT --room $HIPCHAT_ROOM --token $HIPCHAT_TOKEN -s
+    - python slack_notifier.py --org $SLACK_ORG --project $PROJECT --token $SLACK_TOKEN -s
